Refactored the project structure.
CSVReaderUtil:Load the CSV file and map to it

UserBean: Bean class for csv mapping and will be used by DataProvider

moved the testData from src/main/resources to src/test/resources.
